Remove binutils module, no longer needed after microTVM refactor.
